+++ src/main/java/entity/StockMarket.java
@@ -9,12 +9,13 @@ import java.util.concurrent.ConcurrentHashMap;
 /**
  * A singleton class representing the stock market
  */
+// TODO: lock the stock market when updating stock prices
 public class StockMarket {
 
     // thread-safe Singleton instance
     private static volatile StockMarket instance = null;
 
-    private final Map<String, Stock> stocks = new ConcurrentHashMap<>();
+    private Map<String, Stock> stocks = new ConcurrentHashMap<>();
     private IStockDataAccess dataAccess;
     private boolean initialized = false;
 
@@ -48,14 +49,15 @@ public class StockMarket {
     /**
      * Update the stock prices in the stock market
      */
+    // TODO: set interval for updating stock prices
     public void updateStockPrices() {
         if (dataAccess == null) {
             throw new IllegalStateException("StockMarket has not been initialized with a data access object.");
         }
-        Map<String, Double> prices = dataAccess.getStocks();
-        for (Map.Entry<String, Double> entry : prices.entrySet()) {
+        this.stocks = dataAccess.getStocks();
+        for (Map.Entry<String, Stock> entry : stocks.entrySet()) {
             String ticker = entry.getKey();
-            double price = entry.getValue();
+            double price = entry.getValue().getPrice();
             // create a new map entry if stock does not exist, and then update price
             stocks.computeIfAbsent(ticker, k -> new Stock(ticker, price)).updatePrice(price);
         }