+++ src/main/java/interface_adapter/execute_buy/ExecuteBuyController.java
@@ -1,7 +1,9 @@
 package interface_adapter.execute_buy;
 
+import use_case.execute_buy.ExecuteBuyDataAccessInterface;
 import use_case.execute_buy.ExecuteBuyInputBoundary;
 import use_case.execute_buy.ExecuteBuyInputData;
+import utility.ClientSessionManager;
 
 public class ExecuteBuyController {
 
@@ -11,7 +13,12 @@ public class ExecuteBuyController {
         this.interactor = interactor;
     }
 
-    public void execute(ExecuteBuyInputData data) {
+    public void execute(String ticker, String quantity) {
+        final ExecuteBuyInputData data = new ExecuteBuyInputData(
+                ClientSessionManager.Instance().getCredential(),
+                ticker,
+                Integer.parseInt(quantity));
+
         interactor.execute(data);
     }
 }