+++ src/main/java/use_case/execute_buy/ExecuteBuyDataAccessInterface.java
@@ -1,11 +1,9 @@
 package use_case.execute_buy;
 
 import entity.User;
+import utility.exceptions.ValidationException;
 
 public interface ExecuteBuyDataAccessInterface {
     User getUserWithCredential(String credential) throws ValidationException;
 
-    // TODO: define this exception elsewhere
-    class ValidationException extends Exception {
-    }
 }