+++ src/main/java/utility/SessionManager.java
@@ -1,4 +1,4 @@
-package session;
+package utility;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -27,6 +27,7 @@ public class SessionManager {
 
     // create a new session and return the unique credential key
     // TODO: consider hashing the credential key
+    // TODO: consider how to store the token in client-side
     public String createSession(String username) {
         String credentialKey = UUID.randomUUID().toString();
         sessions.put(credentialKey, username);