+++ src/main/java/view/view_events/EventType.java
@@ -1,0 +1,11 @@
+package view.view_events;
+
+
+public enum EventType {
+    SIGN_UP,
+    LOGIN,
+    LOGOUT,
+    UPDATE_ASSET,
+    SWITCH_PANEL,
+    DIALOG,
+}