+++ src/test/java/use_case/execute_buy/ExecuteBuyInteractorTest.java
@@ -11,6 +11,7 @@ import java.util.Optional;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import utility.exceptions.ValidationException;
 
 class ExecuteBuyInteractorTest {
 
@@ -26,12 +27,12 @@ class ExecuteBuyInteractorTest {
     }
 
     @Test
-    void successTest() throws ExecuteBuyDataAccessInterface.ValidationException {
+    void successTest() throws ValidationException {
         User mockUser = userFactory.create("testUser", "password");
         double initialBalance = 10000.0;
         mockUser.addBalance(initialBalance);
 
-        when(dataAccess.getUserWithCredential("00000")).thenReturn(mockUser);
+        when(dataAccess.getUserWithCredential("dummy")).thenReturn(mockUser);
 
         Stock stock = new Stock("XXXX", 100.0);
 
@@ -45,7 +46,7 @@ class ExecuteBuyInteractorTest {
 
             // prepare input data
             int quantityToBuy = 100;
-            ExecuteBuyInputData inputData = new ExecuteBuyInputData("00000", "XXXX", quantityToBuy);
+            ExecuteBuyInputData inputData = new ExecuteBuyInputData("dummy", "XXXX", quantityToBuy);
 
             // create interactor
             ExecuteBuyInteractor interactor = new ExecuteBuyInteractor(dataAccess, outputPresenter);
@@ -72,13 +73,13 @@ class ExecuteBuyInteractorTest {
     }
 
     @Test
-    void insufficientBalanceTest() throws ExecuteBuyDataAccessInterface.ValidationException {
+    void insufficientBalanceTest() throws ValidationException {
         // prepare user with insufficient balance
         User mockUser = userFactory.create("testUser", "password");
         mockUser.addBalance(500.0);
 
         // prepare dataAccess mock
-        when(dataAccess.getUserWithCredential("00000")).thenReturn(mockUser);
+        when(dataAccess.getUserWithCredential("dummy")).thenReturn(mockUser);
 
         // prepare Stock mock
         Stock stock = new Stock("XXXX", 100.0);
@@ -92,7 +93,7 @@ class ExecuteBuyInteractorTest {
             when(stockMarketMock.getStock("XXXX")).thenReturn(Optional.of(stock));
 
             // prepare input data
-            ExecuteBuyInputData inputData = new ExecuteBuyInputData("00000", "XXXX", 100);
+            ExecuteBuyInputData inputData = new ExecuteBuyInputData("dummy", "XXXX", 100);
 
             // create interactor
             ExecuteBuyInteractor interactor = new ExecuteBuyInteractor(dataAccess, outputPresenter);